Use pre-computed footpad state for flywheel ADC checks
Removes ADC_HANDPRESS_SCALE competely. It's unclear how beneficial it
is, but it presents an undocumented and to the user unknown constant
when evaluating ADCs. Meaning the user will set a threshold in the
config, but then, for certain things (Konami) a different threshold is
used that they don't know about.

If there's indeed a need for this, probably a config value should be
added, but a good sensor should activate by a finger's press anyway.

Fix: Remove undocumented 0.8 multiplier for handpress ADC activation.