LEDs: add felony effect
This effect mimics the pattern of common emergency service vehicles.
Its colors are configurable, as well as its speed via Refloat settings.

In the case where the LED strip has an odd number of LEDs, the middle
LED is kept as solid black to ensure the effect is symmetrical.

Feature: Add the Felony LED effect.