Use ticks instead of seconds for timekeeping
The seconds, in float, are converted directly from ticks in the
firmware, meaning when the ticks overflow (which happens roughly every 5
days), the time in float will restart from 0 and all the timers will
fail.

By using ticks directly, as long as we only use relative time intervals,
everything will keep working even with an overflow.

Working with integers is also faster and the switch to ticks saved
almost 200B of flash.

The number of ticks per second is appended to the COMMAND_GET_INFO
message, so that when the timestamp is sent in messages, client apps
know how to convert it to seconds.