Rework the automatic backup feature
The previous approach was unreliable. This new approach adds a flag to
the actual config itself, the flag has a default value of true, but it's
reset to false on every config write.

Still, on the QML side the are no reliable signals that would indicate
the config was read from or written to the board. We now put a hook on
the Refloat Cfg "Write" button to catch that event.

The signalling on the AppUI startup is also unreliable, when the AppUI
and the config are cached, we actually get default config values in the
initial "on loaded" config signal. Some timers and other nonsense is
added to deal with this.